+++ R/get.R
@@ -670,7 +670,7 @@ get_resample_data <- function(qsip_data_object,
 
 #' Get EAF value for a qSIP object
 #'
-#' @param qsip_data_object
+#' @param qsip_data_object a qSIP2 object
 #'
 #' @export
 
@@ -1030,9 +1030,11 @@ summarize_growth_values <- function(qsip_data_object, confidence = 0.9, quiet = 
   is_qsip_growth(qsip_data_object, error = TRUE)
 
   # confirm the confidence value is numeric and between 0-1
-  stopifnot("ERROR: confidence should be numeric" = is.numeric(confidence))
-  if (confidence >= 1 | confidence <= 0) {
-    stop("ERROR: confidence level should be between 0 and 1")
+  if (!is.numeric(confidence)) {
+    cli_abort("{.arg confidence} must be numeric, not {.cls {class(confidence)}}.")
+  }
+  if (confidence <= 0 | confidence >= 1) {
+    cli_abort("{.arg confidence} must be between 0 and 1, not {.val {confidence}}.")
   }
 
   observed <- feature_id <- timepoint1 <- timepoint2 <- N_total_i0 <- N_total_it <- r_net <- bi <- di <- ri <- EAF <- desc <- observed_ri <- NULL